pyproject.toml
[project]
name = "gaitmod"
version = "0.1.0"
description = "A project for analyzing and modeling gait modulation using machine learning techniques."
authors = [{name = "Mohammad Orabe", email = "orabe.mhd@gmail.com"}]
dependencies = [
    "antropy",
    "imbalanced_learn",
    "tensorflow[and-cuda]",
    "matplotlib",
    "mne",
    "numpy",
    "pandas",
    "pyyaml",
    "scikit_learn",
    "scipy",
    "seaborn",
    "setuptools",
    "tqdm",
    "pydot"
]

keywords = ["gait", "modulation", "LFP", "ML"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps."
]

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"